+++ requirements-dev.txt
@@ -7,4 +7,5 @@ pytest-aiohttp==0.3.0
 aiohttp==3.7.3
 yarl==1.6.3
 multidict==5.1.0
+prometheus_client==0.9.0
 backports.zoneinfo;python_version<"3.9"
